requirements.txt
openai==2.17.0
requests==2.32.3
python-dotenv==1.0.1
